Move the code of uploading node headers to the new script
